refactor(format): extract diff rendering module
