+++ ci/proof.toml
@@ -495,6 +495,7 @@ coverage = true
 name = "format_core_outputs"
 kind = "rust"
 paths = [
+  "crates/tokmd-format/src/diff.rs",
   "crates/tokmd-format/src/lib.rs",
   "crates/tokmd-format/tests/format_*.rs",
   "crates/tokmd-format/tests/snapshot*.rs",
